# packages/contracts/openapi.yaml
openapi: 3.1.0
paths:
...
  /incidents/draft-from-text:
    post:
      operationId: IncidentsController_draftFromText
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftIncidentFromTextDto"
      responses:
        "200":
          description: Returns a schema-gated incident draft for staff review.
      tags: &a3
        - incidents
  /incidents:
    post:
      operationId: IncidentsController_create
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateIncidentDto"
      responses:
        "202":
          description: Starts an IncidentReportWorkflow.
      tags: *a3
    get:
      operationId: IncidentsController_list
      parameters: []
      responses:
        "200":
          description: Lists active incidents for the current home.
      tags: *a3
  /incidents/{id}:
    patch:
      operationId: IncidentsController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateIncidentDto"
      responses:
        "202":
          description: ""
      tags: *a3
    get:
      operationId: IncidentsController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Incident with versions and timeline.
      tags: *a3
  /incidents/{id}/submit:
    post:
      operationId: IncidentsController_submit
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitIncidentDto"
      responses:
        "202":
          description: ""
      tags: *a3
  /incidents/{id}/approve:
    post:
      operationId: IncidentsController_approve
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveIncidentDto"
      responses:
        "202":
          description: ""
      tags: *a3
  /incidents/{id}/export:
    post:
      operationId: IncidentsController_export
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportIncidentDto"
      responses:
        "202":
          description: ""
      tags: *a3
  /incidents/{id}/download:
    get:
      operationId: IncidentsController_download
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: Presigned MinIO URL for the exported PDF.
      tags: *a3
  /residents/{id}/timeline:
    get:
      operationId: IncidentsController_residentTimeline
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
      tags: *a3
...
  /assistant/messages:
    post:
      operationId: AssistantController_messages
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssistantMessageDto"
      responses:
        "200":
          description: ""
      tags: *a5
...
info:
  title: CareOS API
  description: Phase 0 CareOS API skeleton.
  version: 0.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    DraftIncidentFromTextDto:
      type: object
      properties:
        free_text:
          type: string
          minLength: 1
          maxLength: 8000
        resident_id:
          type: string
          format: uuid
        template_id:
          type: string
          minLength: 1
          maxLength: 200
      required:
        - free_text
        - resident_id
        - template_id
    CreateIncidentDto:
      type: object
      properties:
        residentId:
          type: string
          format: uuid
        formTemplate:
          type: object
          properties:
            templateId:
              type: string
              minLength: 1
              maxLength: 200
            version:
              type: string
              minLength: 1
              maxLength: 50
          required:
            - templateId
            - version
          x-nestjs_zod-self-required: true
        initialFormData:
          type: object
          additionalProperties: {}
          x-nestjs_zod-self-required: false
      required:
        - residentId
        - formTemplate
    UpdateIncidentDto:
      type: object
      properties:
        formData:
          type: object
          additionalProperties: {}
          x-nestjs_zod-self-required: true
      required:
        - formData
    SubmitIncidentDto:
      type: object
      properties:
        root:
          type: object
          properties: {}
          x-nestjs_zod-unwrap-root: true
          x-nestjs_zod-self-required: false
    ApproveIncidentDto:
      type: object
      properties:
        note:
          type: string
          maxLength: 1000
    ExportIncidentDto:
      type: object
      properties:
        root:
          type: object
          properties: {}
          x-nestjs_zod-unwrap-root: true
          x-nestjs_zod-self-required: false
...
    AssistantMessageDto:
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 4000
        quickActionId:
          type: string
          enum:
            - create_incident
            - notify_safeguarding
            - update_behaviour_log
        residentId:
          type: string
          format: uuid
      required:
        - message
...